# perf: speedup `frappe.qb.get_query` by 2x-4x